+++ lab_3/triangulo.cpp
@@ -1,63 +1,22 @@
-#include <iostream>
-#include <stdexcept>
-#include <bits/stdc++.h>
+#include "triangulo.h"
 
-using namespace std;
 
-class TrianguloRetangulo
+TrianguloRetangulo::TrianguloRetangulo(int a, int b, int c) : lado1(a), lado2(b), lado3(c)
 {
-private:
-    int lado1;
-    int lado2;
-    int lado3;
-
-public:
-    TrianguloRetangulo(int a, int b, int c) : lado1(a), lado2(b), lado3(c)
+    if (!ehTrianguloRetangulo())
     {
-        if (!ehTrianguloRetangulo())
-        {
-            throw std::invalid_argument("Os lados fornecidos nao formam um triangulo valido.");
-        }
+        throw std::invalid_argument("Os lados fornecidos nao formam um triangulo valido.");
     }
+}
 
-    bool ehTrianguloRetangulo() const
-    {
-        vector<int> lados{lado1, lado2, lado3};
-        sort(lados.begin(), lados.end(), greater<int>());
-        return ((lados[0] * lados[0] == lados[2] * lados[2] + lados[1] * lados[1]));
-    }
-
-    void imprimirLados() const
-    {
-        std::cout << "Lados do triangulo retangulo: " << lado1 << ", " << lado2 << ", " << lado3 << std::endl;
-    }
-};
-
-int main()
+bool TrianguloRetangulo::ehTrianguloRetangulo() const
 {
-    int triangulos = 0;
-    for (int a = 1; a < 201; a++)
-    {
-        for (int b = 1; b < 201; b++)
-        {
-            for (int c = 1; c < 201; c++)
-            {
-                try
-                {
-                    // Criar um triângulo retângulo válido
-                    TrianguloRetangulo t1(a, b, c);
-                    t1.imprimirLados();
-                    triangulos++;
-                }
-                catch (const std::invalid_argument &e)
-                {
-                    continue;
-                }
-            }
-        }
-    }
+    vector<int> lados{lado1, lado2, lado3};
+    sort(lados.begin(), lados.end(), greater<int>());
+    return ((lados[0] * lados[0] == lados[2] * lados[2] + lados[1] * lados[1]));
+}
 
-    cout << "\n\n" << triangulos;
-
-    return 0;
+void TrianguloRetangulo::imprimirLados() const
+{
+    std::cout << "Lados do triangulo retangulo: " << lado1 << ", " << lado2 << ", " << lado3 << std::endl;
 }
